All_LLM's/PDF_RAG/requirements.txt
streamlit
embedchain
streamlit-chat
python-dotenv
langchain-huggingface
sentence-transformers